Make PatientController

+++ database/migrations/2021_11_26_091646_create_patients_table.php
@@ -15,7 +15,7 @@ class CreatePatientsTable extends Migration
     {
         Schema::create('patients', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('name');
             $table->string('gender');
             $table->date('date_of_birth');
